% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-bitbucketserver.R
\name{install_bitbucket_server}
\alias{install_bitbucket_server}
\title{Attempts to install a package directly from Bitbucket Dev Server.}
\usage{
install_bitbucket_server(projects, repos,
auth_token = rstudioapi::askForPassword("Enter your personal token:"),
ref = "master", subdir = NULL, quiet = FALSE,
host = NULL, ...)
}
\arguments{
\item{projects}{Project Name}

\item{repos}{Repository Name}

\item{auth_token}{Personal token from Bitbucket Dev Server}

\item{ref}{default: master}

\item{subdir}{default: NULL}

\item{quiet}{default: FALSE}

\item{host}{default: NULL (e.g. https://hostserver.com/bitbucket)}

\item{...}{Other arguments passed on to install.}
}
\description{
This function is vectorised on repo so you can install multiple packages in a single command.
}
\examples{
\dontrun{
install_bitbucket_server(
projects = "project_name",
repos = "repository_name",
host = "host_server_name"
)
}
}
\seealso{
\code{\link[devtools]{install_github}}, \code{\link[devtools]{install_bitbucket}}
}
